updated some routes in routes/api

+++ codeial/controller/api/v1/posts_api.js
@@ -22,19 +22,19 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
         // .id means converting the object id into string
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.deleteOne();
             await Comment.deleteMany({post: req.params.id});
             return res.json(200, {
-                message: "Post and associated comments deleted successfully"
+                message: "Deleted successfully"
             })
-        // }else{
-        //     return res.json(401, {
-        //         message: "You cannot delete this post"
-        //     })
-        // }
+        }else{
+            return res.json(401,{
+                message: "You can't delete this post"
+            })
+        }
     }catch(err){
-        console.log('************Error ***********', err);
+        console.log('************Error(MIGHT COMMENT IS NOT CREATED) ***********', err);
         return res.json(500, {
             message: "Internal Server Error"
         })
